Update README.md
The Credit risk is a possible danger of rejection resulting from a borrower’s failure to refund a loan or meet agreed requirements. Normally, it announces the risk that a lender may not receive the owed principal (actual money) and interest (profit), which results in brake of cash flows and increased prices for setting up (Investopedia, 2017). In this millennium, the banking plays a vital role in our daily life. By using, credit and debit card and online banking, either we are paying or borrowing from banks helps us to fulfil our daily economic needs. The most important part of today’s backing is to provide a landing facility to their customers. Since, the bank provides credit facilities to their customers, they make sure the money is returned and they get some interest (profit) on their actual principal (landed money). It is tough to predict bad loan before landing the money to an applicant. Simply, using Machine learning and innovative machine learning techniques, it has become possible for banks to predict a bad loan application before committing a loan arrangement. Since, the data analytics have played a critical role in modern banking, most banks nowadays are performing credit risk analysis by using the credit score. A loan application will be accepted or rejected based on their result. To execute a credit scoring, a bank uses a model called “Probability of default model” which is also known as PD. By using, the PD model, the bank will decide the price settlement and the cost of capital. In 2004, the world-wide banks agreed to follow some standard rules as their basic criteria for landing.